Created template for home and favourites tabs
