+++ src/components/index.js
@@ -3,3 +3,4 @@ export { default as DetailsAppBar } from './DetailsAppBar';
 export { default as PokemonCard } from './PokemonCard';
 export { default as PokemonGeneralInfo } from './PokemonGeneralInfo';
 export { default as PokemonStats } from './PokemonStats';
+export { default as PokemonsList } from './PokemonsList';
